requirements.txt
pytest
pytest-asyncio
pytest-xdist
playwright
requests
allure-pytest
